@@ -12,7 +12,7 @@
   "dependencies": {
     "core-js": "^3.6.4",
     "register-service-worker": "^1.6.2",
-    "vue": "^2.6.11"
+    "vue": "^3.0.0-alpha.4"
   },
   "devDependencies": {
     "@types/jest": "^24.0.19",
@@ -25,6 +25,7 @@
     "@vue/cli-plugin-typescript": "^4.2.0",
     "@vue/cli-plugin-unit-jest": "^4.2.0",
     "@vue/cli-service": "^4.2.0",
+    "@vue/compiler-sfc": "^3.0.0-alpha.4",
     "@vue/eslint-config-prettier": "^6.0.0",
     "@vue/eslint-config-typescript": "^5.0.1",
     "@vue/test-utils": "1.0.0-beta.31",
@@ -35,6 +36,7 @@
     "sass": "^1.25.0",
     "sass-loader": "^8.0.2",
     "typescript": "~3.7.5",
+    "vue-cli-plugin-vue-next": "^0.0.4",
     "vue-template-compiler": "^2.6.11"
   }
 }
