@@ -1,9 +1,5 @@
-import Vue from "vue";
-import App from "./App.vue";
-import "./registerServiceWorker";
+import { createApp } from 'vue'
+import App from './App.vue'
+import './registerServiceWorker'
 
-Vue.config.productionTip = false;
-
-new Vue({
-  render: h => h(App)
-}).$mount("#app");
+createApp(App).mount('#app')
