@@ -1,20 +1,53 @@
 <template>
   <div id="app">
-    <img alt="Vue logo" src="./assets/logo.png" />
-    <HelloWorld msg="Welcome to Your Vue.js + TypeScript App" />
+    <div>
+      <p
+        v-for="info in messages.state.infos"
+        :key="info"
+        @click="messages.removeInfo(info)"
+      >
+        {{ info }}
+      </p>
+      <p
+        v-for="error in messages.state.errors"
+        :key="error"
+        @click="messages.removeError(error)"
+      >
+        {{ error }}
+      </p>
+    </div>
+    <UploadForm @loadImage="loadImage" />
+    <img :src="state.src" />
   </div>
 </template>
 
 <script lang="ts">
-import Vue from "vue";
-import HelloWorld from "./components/HelloWorld.vue";
+import { reactive } from 'vue'
+import { useMessages } from '@/hooks/messages'
+import UploadForm from '@/components/organisms/UploadForm.vue'
 
-export default Vue.extend({
-  name: "App",
+export default {
+  name: 'App',
   components: {
-    HelloWorld
+    UploadForm
+  },
+  setup() {
+    const state = reactive<{ src: string }>({
+      src: ''
+    })
+    const messages = useMessages()
+    const loadImage = (base64: string) => {
+      console.log(base64)
+      state.src = base64
+    }
+
+    return {
+      state,
+      messages,
+      loadImage
+    }
   }
-});
+}
 </script>
 
 <style lang="scss">
